fix(Backup and Installation script): Fixed a few errors, also issue #25

+++ install.sh
@@ -16,6 +16,9 @@ function is_bookworm() {
 }
 
 function remove_dependencies() {
+	if [ "$test_mode" = true ]; then
+		return
+	fi
 	if is_bookworm; then
 		pip3 uninstall telegram python-telegram-bot --break-system-packages
 	else
@@ -24,6 +27,9 @@ function remove_dependencies() {
 }
 
 function install_dependencies() {
+	if [ "$test_mode" = true ]; then
+		return
+	fi
 	if is_bookworm; then
 		pip3 install python-telegram-bot==13.15 --break-system-packages
 	else
@@ -89,22 +95,43 @@ function get_installation_path() {
 	if [ -z "${installation_dir//\"/}" ] || [ "$installation_dir" = true ]; then
 		echo "[ ! ] The installation directory was not found in the configuration file"
 		read -r -p "Please enter the installation directory, press Enter to set '/usr/local/share/pwnagotchi/custom-plugins' or specify yours with absolute path: " installation_dir
+		if [ -z "${installation_dir//\"/}" ]; then
+			installation_dir="/usr/local/share/pwnagotchi/custom-plugins"
+		fi
+		edit_configuration_values "main.custom_plugins" "${installation_dir}" "$CONFIG_FILE"
+		installation_dir="${installation_dir//\"/}"
 	fi
-	if [ -z "${installation_dir//\"/}" ]; then
-		installation_dir="/usr/local/share/pwnagotchi/custom-plugins"
-	fi
-	edit_configuration_values "main.custom_plugins" "${installation_dir}" "$CONFIG_FILE"
-	installation_dir="${installation_dir//\"/}"
 }
 
 # Main
 
 # Check that the script is running as root
+# If the -t flag is used, the script will be run in test mode
+# In test mode, the script will not need to run as root, and will not perform the remove and installation of dependencies.
 
-if [ "$EUID" -ne 0 ]; then
+# Capture user arguments
+
+while getopts "t" options; do
+	case $options in
+	t)
+		test_mode=true
+		;;
+	\?)
+		echo "Invalid option: -$OPTARG" >&2
+		exit 1
+		;;
+	esac
+done
+
+if [ "$test_mode" = true ]; then
+	echo "[ * ] Running in test mode"
+	echo "[ * ] Dependencies will not be removed or installed"
+	CONFIG_FILE="config.toml"
+elif [ "$EUID" -ne 0 ]; then
 	echo "[ ! ] This script need to be run as root"
 	exit 0
 fi
+
 user_sleep
 echo "[ + ] Getting installation path..."
 get_installation_path
